+++ src/SFA.DAS.RequestApprenticeshipTraining.UITests/Project/Helpers/EmployerPortalViaRatLoginHelper.cs
@@ -4,17 +4,16 @@ using SFA.DAS.Registration.UITests.Project.Tests.Pages.StubPages;
 using SFA.DAS.RequestApprenticeshipTraining.UITests.Project.Tests.Pages.RATEmployerPages;
 using TechTalk.SpecFlow;
 
-namespace SFA.DAS.RequestApprenticeshipTraining.UITests.Project.Helpers
+namespace SFA.DAS.RequestApprenticeshipTraining.UITests.Project.Helpers;
+
+public class EmployerPortalViaRatLoginHelper(ScenarioContext context) : EmployerPortalLoginHelper(context)
 {
-    public class EmployerPortalViaRatLoginHelper(ScenarioContext context) : EmployerPortalLoginHelper(context)
+    public AskIfTrainingProvidersCanRunThisCoursePage LoginViaRat(RatEmployerBaseUser loginUser)
     {
-        public AskIfTrainingProvidersCanRunThisCoursePage LoginViaRat(RatEmployerBaseUser loginUser)
-        {
-            SetCredentials(loginUser, true);
+        SetCredentials(loginUser, true);
 
-            new StubSignInEmployerPage(context).Login(loginUser).Continue();
+        new StubSignInEmployerPage(context).Login(loginUser).Continue();
 
-            return new(context);
-        }
+        return new(context);
     }
 }
