+++ src/SFA.DAS.RequestApprenticeshipTraining.UITests/Project/Tests/Pages/RATEmployerPages/AskIfTrainingProvidersCanRunThisCoursePage.cs
@@ -1,21 +1,20 @@
 ﻿using OpenQA.Selenium;
 using TechTalk.SpecFlow;
 
-namespace SFA.DAS.RequestApprenticeshipTraining.UITests.Project.Tests.Pages.RATEmployerPages
+namespace SFA.DAS.RequestApprenticeshipTraining.UITests.Project.Tests.Pages.RATEmployerPages;
+
+public class AskIfTrainingProvidersCanRunThisCoursePage(ScenarioContext context) : RatProjectBasePage(context)
 {
-    public class AskIfTrainingProvidersCanRunThisCoursePage(ScenarioContext context) : RatProjectBasePage(context)
+    protected override string PageTitle => "Ask if training providers can run this course";
+
+    #region Locators
+    private static By ClickStartNowButton => By.CssSelector("a.govuk-button.govuk-button--start");
+    #endregion
+
+    public HowManyAprenticesWouldDoThisApprenticeshipTrainingPage ClickStarNow()
     {
-        protected override string PageTitle => "Ask if training providers can run this course";
+        formCompletionHelper.Click(ClickStartNowButton);
 
-        #region Locators
-        private static By ClickStartNowButton => By.CssSelector("a.govuk-button.govuk-button--start");
-        #endregion
-
-        public HowManyAprenticesWouldDoThisApprenticeshipTrainingPage ClickStarNow()
-        {
-            formCompletionHelper.Click(ClickStartNowButton);
-
-            return new HowManyAprenticesWouldDoThisApprenticeshipTrainingPage(context);
-        }
+        return new HowManyAprenticesWouldDoThisApprenticeshipTrainingPage(context);
     }
 }
