+++ src/SFA.DAS.RequestApprenticeshipTraining.UITests/Project/Tests/Pages/RATEmployerPages/RatProjectBasePage.cs
@@ -2,12 +2,11 @@
 using SFA.DAS.UI.Framework.TestSupport;
 using TechTalk.SpecFlow;
 
-namespace SFA.DAS.RequestApprenticeshipTraining.UITests.Project.Tests.Pages.RATEmployerPages
+namespace SFA.DAS.RequestApprenticeshipTraining.UITests.Project.Tests.Pages.RATEmployerPages;
+
+public abstract class RatProjectBasePage : VerifyBasePage
 {
-    public abstract class RatProjectBasePage : VerifyBasePage
-    {
-        protected RatProjectBasePage(ScenarioContext context, bool verifyPage = true) : base(context) { if (verifyPage) VerifyPage(); }
-        
-        protected override By ContinueButton => By.CssSelector("[id='main-content'] button.govuk-button[type='submit']");
-    }
+    protected RatProjectBasePage(ScenarioContext context, bool verifyPage = true) : base(context) { if (verifyPage) VerifyPage(); }
+    
+    protected override By ContinueButton => By.CssSelector("[id='main-content'] button.govuk-button[type='submit']");
 }
